Made the code more succinct and modular by diving into header, functions and main file
